build(frontend): don't mount frontend in container
HACK: `npm install` wasn't executed in the container, failing the build. The
volume seems to be the culprit, instead copy the whole folder. This
results in reproducible builds, but disables the hot reloading of the
dev server. You can run the backend independently and use a local dev
server if you want for now.