build(frontend): disable frontend container
The Quasar dev server needs a different setup to run from inside a
container, as it has it's own CLI and there is no `--host` option.
Since the container is broken, the service is disable in
`docker-compose.yml` and the frontend `Dockerfile` removed. This way
docker compose won't choke on the frontend service and you don't need to
specify that you only want the backend.