fix(user): limit permissions to user, set password
Previously it was possible for a logged in user to view, edit or delete
other user's records. A custom permission class was added to adress this.
Updating the user's password would not use the correct implementation
and store a raw password instead of salting and hashing. This resulted
in being unable to login after changing the password.
It was fixed by overwriting the update method of the Serializer.
Includes test cases for the described behaviour.

Fix: #5