feat(tokens): rotate refresh tokens
Using a refresh token will return a new refresh token in addition to an
access token. That way a user session can be extended automatically, when
used before the original refresh token expires. Previously the only way
to obtain a new refresh token was by authenticating again through
`/api/users/token/`
That made it necessary to store user credentials in the frontend
to refresh the token or requiring a new login every 24 hours regardless
of activity.
Refresh tokens are now valid for 3 days, access tokens for 5 minutes.

Fix: #3