feat(user)!: split user and profile
Introduces User profiles (1-to-1) and a preliminary API to query them.
Profiles are linked through the pk field "user", which corresponds to
the users "id" field. A read-only hyperlinked profile url is included in
the users API, to facilitate fetching the profile on a successful login.
Since the User implementation was renamed from CustomUser to User, the
initial migration is inconsistent and needs a database reset. The
following migrations removed to keep the migration history useful.

BREAKING CHANGE: requires a fresh database