refactor(users)!: split into cycles, notes
Refactored users into different django apps.
Cycles holds everything related to menstruation tracking and phases.
Notes are miscellaneous daily tracking entries that are not relevant for
the calculations or training recommendations. Training endpoints should
go to a new training app.
There are also changes to filtering and a users.views.UserModelViewSet
as an abstraction to pre-filter querysets for the logged-in user. Tests
are included for notes. This part is more than a refactoring.

BREAKING CHANGE: Trailing slash for all routes