fix(users): permissions for user objects
The custom permission for user objects only worked if the access object
was a user object, i.e. /api/users/.
Therefore it was not possible to access other user related endpoints,
e.g. the Profile.
Permissions were changed to also allow access to objects that contain
the user as foreign key. This is the case for almost all endpoints. The
field name is expected to be `user`.