test(profile): readonly pk user
Changes the user primary and foreign key of the Profile to read-only,
so it can not be accidentally or maliciously be updated for a wrong user.
Tests indicated that it was not possible before, as there is a one to
one relationship between users and profiles and there would be an error
trying to add another profile to an existing user or non-existent user.
Since the values has no meaning anyways, making it read-only reflects
the usage. This is not a breaking change, as the endpoint will accept
arbitrary keys and ignore them. So pre-existing requests that include a
user will still work and not change their behaviour if they worked before.
Adds tests for those cases and documentation for the usage of patch for
partial updates and how to set optional profile settings to null.